preprocessing/requirements.txt
opencv-python
pillow
wget